requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
pyyaml>=6.0.0
xgboost>=2.0.0
shap>=0.43.0
kaggle
dvc
mlflow
boto3>=1.34.0
s3fs>=2024.2.0
dvc-s3>=3.0.1